daemon: start proxy before restoring old endpoints
Each endpoint now has the daemon's proxy injected into it for callbacks into the
proxy during regeneration. So, on daemon bootstrap, when we restore old
endpoints, we need to inject the daemon's proxy into each restored endpoint. The
only way we can do that is if the daemon's proxy is initialized before we
restore endpoints. Do so.

Signed-off by: Ian Vernon <ian@cilium.io>