daemon: do not remove created directory in `TearDownTest`
The directory is created in `TestMain`, which runs once, while `TearDownTest`
runs after every test. This means that the directory is deleted after the first
test in the Daemon test suite is ran, which is not correct. Instead, defer the
deletion of this directory after `TestMain` has exited.

Signed-off by: Ian Vernon <ian@cilium.io>